% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractSymMat.R
\docType{class}
\name{abstractSymMat}
\alias{abstractSymMat}
\title{Abstract Class: \code{AbstractSymMat}}
\arguments{
\item{x, object}{a \code{AbrastSymMat} object}
}
\description{
A class handling the storage of a symmetric square matrix.
}
\section{Fields}{

\describe{
\item{\code{dim}}{The dimension of the square matrix (numeric).}

\item{\code{names}}{A vector storing the names for rows and columns (vector).}

\item{\code{defaultDiag}}{The default values for diagonal elements
(numericOrLogicalOrNULL).}
}}


\keyword{internal}
