% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractSymMat.R, R/symMMat.R, R/symVMat.R
\name{abstractSymMat_access}
\alias{abstractSymMat_access}
\alias{[,AbstractSymMat,numeric,missing,ANY-method}
\alias{[,AbstractSymMat,missing,numeric,ANY-method}
\alias{[,AbstractSymMat,missing,missing,ANY-method}
\alias{[,SymMMat,matrix,missing,ANY-method}
\alias{[,SymMMat,numeric,numeric,ANY-method}
\alias{[,SymVMat,matrix,missing,ANY-method}
\alias{[,SymVMat,numeric,numeric,ANY-method}
\title{Access to \code{AbstractSymMat} data`}
\usage{
\S4method{[}{AbstractSymMat,numeric,missing,ANY}(x, i, j, drop)

\S4method{[}{AbstractSymMat,missing,numeric,ANY}(x, i, j, drop)

\S4method{[}{AbstractSymMat,missing,missing,ANY}(x, i, j, drop)

\S4method{[}{SymMMat,matrix,missing,ANY}(x, i, j, drop)

\S4method{[}{SymMMat,numeric,numeric,ANY}(x, i, j, drop)

\S4method{[}{SymVMat,matrix,missing,ANY}(x, i, j, drop)

\S4method{[}{SymVMat,numeric,numeric,ANY}(x, i, j, drop)
}
\arguments{
\item{x}{An \code{AbstractSymMat} or inheriting object}

\item{i}{A vector composed of the indices of the wished rows.}

\item{j}{A vector composed of the indices of the wished columns}
}
\value{
Depending on \code{drop} and the dimensions of \code{i} and \code{j}, a matrix or
a vector.
}
\description{
Rules are the following:
\itemize{
\item when rows and columns are given a submatrix is returned with the
asked rows and columns
\item when only rows (resp. columns are given) a submatrix is return
with the asked rows (resp. columns) and the entire columns (resp. rows)
\item When rows and columns are missing the entire matrix is sent back.
Depending on \code{drop} (a flag) is there is only one row and / or one
column given the result will be simplified in a vector form or not.
}
}
\keyword{internal}
