% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractSymMat.R, R/distMat.R
\name{abstractSymMat_properties}
\alias{abstractSymMat_properties}
\alias{names,AbstractSymMat-method}
\alias{length,AbstractSymMat-method}
\alias{dim,AbstractSymMat-method}
\alias{nrow,AbstractSymMat-method}
\alias{ncol,AbstractSymMat-method}
\alias{isSymmetric,AbstractSymMat-method}
\alias{is.matrix,AbstractSymMat-method}
\alias{names,DistMat-method}
\alias{length,DistMat-method}
\alias{dim,DistMat-method}
\alias{nrow,DistMat-method}
\alias{ncol,DistMat-method}
\alias{isSymmetric,DistMat-method}
\alias{is.matrix,DistMat-method}
\alias{is.numeric,DistMat-method}
\title{Access properties of a \code{AbstractSymMat} object or subclasses}
\usage{
\S4method{names}{AbstractSymMat}(x)

\S4method{length}{AbstractSymMat}(x)

\S4method{dim}{AbstractSymMat}(x)

\S4method{nrow}{AbstractSymMat}(x)

\S4method{ncol}{AbstractSymMat}(x)

\S4method{isSymmetric}{AbstractSymMat}(object, ...)

\S4method{is.matrix}{AbstractSymMat}(x)

\S4method{names}{DistMat}(x)

\S4method{length}{DistMat}(x)

\S4method{dim}{DistMat}(x)

\S4method{nrow}{DistMat}(x)

\S4method{ncol}{DistMat}(x)

\S4method{isSymmetric}{DistMat}(object, ...)

\S4method{is.matrix}{DistMat}(x)

\S4method{is.numeric}{DistMat}(x)
}
\arguments{
\item{x}{An \code{AbstractSymMat} object.}
}
\value{
For \code{isSymmetric} : always TRUE as any \code{AbstractSymMat}
is symmetric.

For \code{is.matrix} : always \code{TRUE} as any \code{AbstractSymMat} is a matrix.
}
\description{
Access properties of a \code{AbstractSymMat} object or subclasses

integers vector)
}
\keyword{internal}
