% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractSymMat.R, R/distMat.R, R/symVMat.R
\name{abstractSymMat_replace_diag}
\alias{abstractSymMat_replace_diag}
\alias{diag<-,DistMat-method}
\alias{diag<-,SymVMat-method}
\title{Replace diagonal values of a \code{AbstractSymMat} object or subclasses}
\usage{
\S4method{diag}{DistMat}(x) <- value

\S4method{diag}{SymVMat}(x) <- value
}
\arguments{
\item{x}{a \code{AbstractSymMat} object or subclasses}

\item{value}{a 1 or \code{nrow(x)} length vector}
}
\description{
Replace diagonal values of a \code{AbstractSymMat} object or subclasses
}
\keyword{internal}
