% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractSymMat.R
\name{abstractSymMat_replace_name}
\alias{abstractSymMat_replace_name}
\alias{names<-,AbstractSymMat,vector-method}
\title{Replace the names of elements in an \code{AbstractSymMat} or subclasses}
\usage{
\S4method{names}{AbstractSymMat,vector}(x) <- value
}
\arguments{
\item{x}{An \code{AbstractSymMat} (or subclasses) object.}

\item{value}{A vector of names to be assigned to rows and columns.}
}
\value{
The updated object.
}
\description{
Replace the names of elements in an \code{AbstractSymMat} or subclasses
}
\keyword{internal}
