% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHCTree.R
\name{AHCTree_access}
\alias{AHCTree_access}
\alias{[,AHCTree,logical,ANY,ANY-method}
\alias{[,AHCTree,vector,ANY,ANY-method}
\alias{[[,AHCTree,numeric-method}
\alias{[[,AHCTree,character-method}
\alias{[[,AHCTree,missing-method}
\title{Access to the partitions of a \code{AHCTree}}
\usage{
\S4method{[}{AHCTree,logical,ANY,ANY}(x, i)

\S4method{[}{AHCTree,vector,ANY,ANY}(x, i)

\S4method{[[}{AHCTree,numeric}(x, i)

\S4method{[[}{AHCTree,character}(x, i)

\S4method{[[}{AHCTree,missing}(x, i)
}
\arguments{
\item{x}{A \code{AHCTree} object}

\item{i}{As logical, strictly positive integers or character
vector.}
}
\value{
for "[", a list of \code{Partition} objects ; for "[[" a \code{Partition}
object if \code{i} is of length 1, a list of \code{Partition} objects otherwise.
}
\description{
Access to the partitions of a \code{AHCTree}
}
\keyword{internal}
