% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHCTree.R
\name{AHCTree_check_constraints}
\alias{AHCTree_check_constraints}
\title{Does partitions of a \code{AHCTree} verifies constraints?}
\arguments{
\item{x}{A \code{AHCTree}}
}
\description{
Does partitions of a \code{AHCTree} verifies constraints?
}
\keyword{internal}
