% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHCTree.R
\name{AHCTree_constructor}
\alias{AHCTree_constructor}
\title{Create an \code{AHCTree} object.}
\usage{
AHCTree_constructor(pb, partitions = list(), n = NULL)
}
\arguments{
\item{pb}{An object of class \code{pbCon} representing the clustering problem.}

\item{partitions}{A list of Partition objects or a single Partition object.}

\item{n}{Number of partitions to create if a single Partition object
is provided. Default is NULL.}
}
\value{
An \code{AHCTree} object with the specified clustering problem
and partitions.
}
\description{
This function creates an \code{AHCTree} object with the given clustering problem
and partitions. The partitions can be provided as a list of Partition
objects or as a single Partition object with the number of desired partitions
(n). If n is specified and greater than 1, it will create
n copies of the given Partition object.
}
\keyword{internal}
