% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractSymMat.R
\name{anyNA,AbstractSymMat-method}
\alias{anyNA,AbstractSymMat-method}
\title{Is there any NA is our symmetric matrix?}
\usage{
\S4method{anyNA}{AbstractSymMat}(x)
}
\arguments{
\item{x}{An \code{AbstractSymMat} object.}
}
\value{
Logical value indicating if any NA values are present in the
\code{AbstractSymMat} object.
}
\description{
Is there any NA is our symmetric matrix?
}
\keyword{internal}
