% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat.R
\name{arguments_distMat}
\alias{arguments_distMat}
\alias{constructor_DistMat}
\title{Constructor function for creating a \code{DistMat} object}
\usage{
constructor_DistMat(
  distances = NULL,
  d = NULL,
  data = NULL,
  standardQuant = FALSE,
  binarQual = FALSE,
  storageMode = "vector",
  names = NULL,
  p = 2L
)
}
\arguments{
\item{distances}{The distance matrix of the problem. This can be omitted
if a distance function \code{d} and data context \code{data} are provided. If only
\code{distances} is provided, all distances must be present. (distance matrix)}

\item{d}{Distance function between elements. This can be omitted if
\code{distances} is already indicated. If present, \code{data} must also be
specified. Some classical distances are available, it is recommended to use
them rather than a personal function for optimisation reasons :
\itemize{
\item "\code{euclidean}": Euclidean distance.
\item "\code{manhattan}" : Manhattan distance.
\item "\code{minkowski}" : Minkowski's distance. In that case a value for p >= 1
must be specified.
}

(function or string)}

\item{data}{A data.frame where each row represents data related to an
element. This can be omitted if \code{d} is omitted. Present variables can
be quantitative or qualitative. If qualitative variables are present,
some distances may not be used. Possibility of standardising variables and
transforming qualitative variables into binary variables (one-hot encoding)
using \code{standardQuant} and \code{binarQual}. (data.frame)}

\item{standardQuant}{\code{TRUE} if the variables in \code{data} should be
standardised (i.e., centered and scaled), \code{FALSE} (default) otherwise.
Standardisation is applied after the possible binarization of qualitative
variables (see \code{binarQual}). (flag)}

\item{binarQual}{\code{TRUE} if qualitative variables should be binarized (one-hot
encoding), for example, to make the data set compatible with common distances
or to standardize these variables. \code{FALSE} (default) otherwise. (flag)}

\item{storageMode}{Indicates how distance data should be stored. Do not
modify for now. (string)}

\item{names}{A character vector with names for the elements in
the distance matrix.}

\item{p}{An integer value for the Minkowski distance parameter.}
}
\value{
A \code{DistMat} object representing the distance matrix.
}
\description{
Constructor function for creating a \code{DistMat} object
}
\keyword{internal}
