% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{available_criteria}
\alias{available_criteria}
\title{Available clustering criteria}
\usage{
available_criteria()
}
\description{
Give what criteria are available actually.
}
\details{
Silhouette score will be available if \code{cluster} package
is installed.
}
\examples{
available_criteria()
}
\seealso{
Other available parameters: 
\code{\link{available_fusion_constraints}()},
\code{\link{available_linkages}()}
}
\concept{available parameters}
