% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_elements.R
\name{available_elements_distances}
\alias{available_elements_distances}
\title{What elements distances are already implemented?}
\usage{
available_elements_distances()
}
\description{
What elements distances are already implemented?
}
\seealso{
Other available parameters: 
\code{\link{available_criteria}()},
\code{\link{available_fusion_constraints}()},
\code{\link{available_linkages}()}
}
\concept{available parameters}
\keyword{internal}
