% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_constraint.R
\name{available_fusion_constraints}
\alias{available_fusion_constraints}
\alias{available_fusion_modes}
\title{Available fusion constraints}
\usage{
available_fusion_constraints()

available_fusion_modes()
}
\value{
a character vector with available choices.
}
\description{
What parameters are available for fusion constraints?
}
\section{Functions}{
\itemize{
\item \code{available_fusion_constraints()}: return implemented
fusion constraints.

\item \code{available_fusion_modes()}: return implemented
fusion constraint modes.

}}
\seealso{
Other available parameters: 
\code{\link{available_criteria}()},
\code{\link{available_elements_distances}()},
\code{\link{available_linkages}()}

Other available parameters: 
\code{\link{available_criteria}()},
\code{\link{available_elements_distances}()},
\code{\link{available_linkages}()}
}
\concept{available parameters}
\keyword{internal}
