% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_distances.R
\name{available_linkages}
\alias{available_linkages}
\title{What linkages are available?}
\usage{
available_linkages(description = FALSE)
}
\arguments{
\item{description}{flag indicating if description of implemented
linkages should be printed.}
}
\value{
A character vector containing names of implemented linkages.
If \code{description = TRUE} return will me bade invisibly.
}
\description{
Give the list of implemented linkages.
}
\seealso{
Other available parameters: 
\code{\link{available_criteria}()},
\code{\link{available_fusion_constraints}()}
}
\concept{available parameters}
