% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calinski_harabasz.R
\name{calinski_harabasz}
\alias{calinski_harabasz}
\title{Calinski-Harabasz Index (CHI / ICH)}
\usage{
calinski_harabasz(data, partition, valueOnly = TRUE)
}
\arguments{
\item{data}{A data frame with one row per element, containing context.
Each variable included must be compatible with the euclidean distance
(and so have to be quantitative) because CHI use the variance.}

\item{partition}{The partition of the problem: a vector with a length
equal to the number of rows of \code{data}.}

\item{valueOnly}{\code{TRUE} if the only value needed is the index. \code{FALSE}
if the user wants more information.}
}
\value{
Depends on \code{valueOnly}.
If \code{valueOnly = TRUE}, returns only the Calinski-Harabasz index
(positive double value).
If \code{valueOnly = FALSE}, returns a list containing the following elements:
\itemize{
\item "\code{k}": number of clusters in the partition (strictly positive integer)
\item "\code{nbElementsClusters}": strictly positive integers vector of size \code{k}
giving the number of elements of each cluster.
\item "\code{withinClusters}": A positive double vector of size \code{k} containing
the within sum of squares of each cluster.
\item "\code{W}": The within sum of squares of the partition, i.e. the sum of the
empirical cluster variances weighted by the number of elements of each
clusters. (positive real value)
\item "\code{betweenClusters}": A positive double vector of size \code{k} containing for
each cluster the square euclidean norm between their and the global centroid
(i.e. ||c_i - c||²)
\item "\code{B}" : The between sum of squares of the partition, i.e. the sum of the
square euclidean norm given in "\code{betweenClusters}" weighted by the number
of elements of each cluster.
\item "\code{CHI}" : The Calinski-Harabasz Index (positive real value)
}
}
\description{
Methods to determine the value of the Calinski-Harabasz Index of a
partition with elements variable(s) compatible with euclidean distance.
This index is defined by between variance divided by within variance,
each divided by their degree of liberty.
Depending of \code{valueOnly}, the value returned by \code{calinski_harabasz} if
only the index or a list with more data. It can be useful for instance to
update the index after some change instead of start from the beginning.
}
\details{
Some special cases can be considered:
\itemize{
\item \code{n = 1} : their is only one possible partition so it is optimal by
definition. Hence we take \code{CHI = Inf}.
\item \code{k = 1}: the unique cluster of the partition is the entire set.
Here the between sum of squares (B) is considered equal to 0. In the index
formula \code{B} is divided by \code{k-1} which equals \code{0} here. Except when \code{n = 1}
this partition is of no interest so we assign \code{CHI = 0}.
\item \code{k = n} : each cluster is made of one element. Consequently the empirical
biased variance of each cluster is equal to 0, so is \code{W}. In the index
formula \code{W} is divided by \code{n - k} which is equal to 0 here. In that case
compactness is optimal but separation is at its worth. The index is set to
\code{NaN}.
}
}
\references{
Calinski, T., and Harabasz, J. (1974) A Dendrite Method for
Cluster Analysis, Communications in Statistics, 3, 1-27.
}
\seealso{
\code{calinhara()} from package
\href{https://CRAN.R-project.org/package=fpc}{fpc}
which calculates the CHI and return only its value. This function can be 2
or 3 times faster than \code{calinski_harabasz} but take a lot more of memory
(empirically the ratio "memory \code{fpc}" / "memory \code{c3t}" seems to increase
linearly with \code{n}, being around 80 at \code{n = 3500}).
}
\keyword{internal}
