% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_tools.R, R/contiguity.R
\name{is_feasible_solution}
\alias{is_feasible_solution}
\alias{check_solution}
\alias{is_regionalisation}
\title{Is a partition checking constraints?}
\usage{
is_feasible_solution(
  partition,
  contiguity = NULL,
  sizes = rep(1L, length(partition)),
  m = 0,
  M = Inf
)

is_regionalisation(partition, contiguity)
}
\arguments{
\item{partition}{Partition vector of the set (numeric vector
composed of non-null integers).}

\item{contiguity}{contiguity matrix or graph. If \code{NULL} then the contiguity
constraint is not considered.}

\item{sizes}{sizes of the elements (positive values)}

\item{m, M}{minimum and maximum size constraints. (positive values)}
}
\value{
A flag equals to \code{TRUE} if \code{partition} check the requested
constraints, \code{FALSE} otherwise.
}
\description{
Is a partition checking constraints?
}
\section{Functions}{
\itemize{
\item \code{is_feasible_solution()}: Verifies if a partition check all connectivity
and size constraints

\item \code{is_regionalisation()}: Verifies if all classes in a set are connected.

}}
