% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{checkCharacterSubset}
\alias{checkCharacterSubset}
\alias{assertCharacterSubset}
\alias{testCharacterSubset}
\alias{checkStringChoice}
\alias{assertStringChoice}
\alias{testStringChoice}
\title{Check if an argument is a subset of a character vector}
\usage{
checkCharacterSubset(
  x,
  choices,
  keepCase = FALSE,
  keepAccents = FALSE,
  len = NULL,
  empty.ok = FALSE
)

assertCharacterSubset(
  x,
  choices,
  keepCase = FALSE,
  keepAccents = FALSE,
  len = NULL,
  empty.ok = FALSE,
  .var.name = checkmate::vname(x),
  add = NULL
)

testCharacterSubset(
  x,
  choices,
  keepCase = FALSE,
  keepAccents = FALSE,
  len = NULL,
  empty.ok = FALSE
)

checkStringChoice(x, choices, keepCase = FALSE, keepAccents = FALSE)

assertStringChoice(
  x,
  choices,
  keepCase = FALSE,
  keepAccents = FALSE,
  .var.name = checkmate::vname(x),
  add = NULL
)

testStringChoice(x, choices, keepCase = FALSE, keepAccents = FALSE)
}
\arguments{
\item{x}{The argument to be tested.}

\item{choices}{The non-empty character vectors \code{x} is supposed to
come.}

\item{keepCase}{\code{TRUE} if case must be maintained. If \code{FALSE} (default)
each letter in \code{x} strings are lowered, as well as those in \code{choices}.}

\item{keepAccents}{\code{TRUE} if accents must be maintained. If \code{FALSE}
(default) accents are removed in \code{x} and \code{choices} strings.}

\item{len}{integer indicating the size \code{x} is supposed to have.
Default to \code{NULL} (do verification made).}

\item{.var.name}{Name of the checked object to print in assertions.
Defaults to the heuristic implemented in \code{\link[checkmate:vname]{checkmate::vname()}}.}
}
\value{
Please see \code{\link[checkmate:checkMatrix]{checkmate::checkMatrix()}} help page (for instance)
for further details about the difference between check_, test_ and assert_
functions results.
}
\description{
Check if an argument is a subset of a character vector
}
\details{
Main differences with \code{\link[checkmate:checkSubset]{checkmate::checkSubset()}} is that
it specific to characters and allow before comparison to decide
if case and accents must be considered.
}
\section{Functions}{
\itemize{
\item \code{checkStringChoice()}: Shortcut for when \code{len = 1L},
i.e. \code{x} must be one element of \code{choices}.

}}
\keyword{internal}
