% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{checkContiguityGraph}
\alias{checkContiguityGraph}
\alias{assertContiguityGraph}
\alias{testContiguityGraph}
\title{Check if an argument is a contiguity graph}
\usage{
checkContiguityGraph(g, n = gorder(g))

assertContiguityGraph(
  g,
  n = gorder(g),
  .var.name = checkmate::vname(g),
  add = NULL
)

testContiguityGraph(g, n = gorder(g))
}
\arguments{
\item{g}{The argument to try.}

\item{n}{The order (number of vertices) \code{g} is supposed to have. By
default this number is arbitrary.}

\item{.var.name}{Name of the checked object to print in assertions.
Defaults to the heuristic implemented in \code{\link[checkmate:vname]{checkmate::vname()}}.}

\item{add}{Collection to store assertion messages.
See \link[checkmate:AssertCollection]{checkmate::AssertCollection}.}
}
\value{
Please see \code{\link[checkmate:checkMatrix]{checkmate::checkMatrix()}} help page (for instance)
for further details about the difference between check_, test_ and assert_
functions results.
}
\description{
Check if an argument is a \code{ìgraph} object from \code{igraph} package
with contiguity graph properties, i.e. being simple and undirected.
}
\details{
Extension of the \code{checkmate} package.
}
\examples{
require(igraph)

graph <- make_graph(c(1, 2, 2, 3, 3, 1), directed = FALSE)

checkContiguityGraph(graph) # TRUE
}
\seealso{
Other arguments checkers: 
\code{\link{checkContiguityMatrix}()},
\code{\link{checkDistanceMatrix}()},
\code{\link{checkPartition}()},
\code{\link{checkRegionalisation}()}
}
\concept{arguments checkers}
\keyword{internal}
