% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{checkContiguityMatrix}
\alias{checkContiguityMatrix}
\alias{assertContiguityMatrix}
\alias{testContiguityMatrix}
\title{Check if an argument is a contiguity matrix}
\usage{
checkContiguityMatrix(
  x,
  isComplete = TRUE,
  any.missing = FALSE,
  all.missing = FALSE,
  nrows = NULL,
  ncols = nrows
)

assertContiguityMatrix(
  x,
  isComplete = TRUE,
  any.missing = FALSE,
  all.missing = FALSE,
  nrows = NULL,
  ncols = nrows,
  .var.name = checkmate::vname(x),
  add = NULL
)

testContiguityMatrix(
  x,
  isComplete = TRUE,
  any.missing = FALSE,
  all.missing = FALSE,
  nrows = NULL,
  ncols = nrows
)
}
\arguments{
\item{x}{Matrix to be tested}

\item{isComplete}{A flag set to \code{TRUE} if \code{x} is supposed to be
a distance matrix of a complete set, i.e. must be symmetric with 0
at the diagonal (can be NA if \code{any.missing = TRUE}).}

\item{any.missing}{Are missing values allowed? Default is \code{FALSE}.}

\item{all.missing}{Are matrices with only missing values allowed?
Default is \code{FALSE}.}

\item{nrows, ncols}{exact number of rows and columns. If \code{NULL} check is
ignored. If \code{isComplete} is \code{TRUE} then \code{nrows} must be equals to \code{ncols}}

\item{.var.name}{Name of the checked object to print in assertions.
Defaults to the heuristic implemented in \code{\link[checkmate:vname]{checkmate::vname()}}.}

\item{add}{Collection to store assertion messages.
See \link[checkmate:AssertCollection]{checkmate::AssertCollection}.}
}
\value{
Please see \code{\link[checkmate:checkMatrix]{checkmate::checkMatrix()}} help page (for instance)
for further details about the difference between check_, test_ and assert_
functions results.
}
\description{
Check if an argument is a matrix with contiguity matrix properties.
}
\details{
Extension of the \code{checkmate} package.
}
\seealso{
Other arguments checkers: 
\code{\link{checkContiguityGraph}()},
\code{\link{checkDistanceMatrix}()},
\code{\link{checkPartition}()},
\code{\link{checkRegionalisation}()}
}
\concept{arguments checkers}
