% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{checkCriteria}
\alias{checkCriteria}
\alias{assertCriteria}
\alias{testCriteria}
\alias{checkCriterion}
\alias{assertCriterion}
\alias{testCriterion}
\alias{checkCompatibleCriteria}
\alias{assertCompatibleCriteria}
\alias{testCompatibleCriteria}
\alias{checkCompatibleCriterion}
\alias{assertCompatibleCriterion}
\alias{testCompatibleCriterion}
\title{Checking is an argument is a criterion
Check if an argument is a correct clustering criterion.}
\usage{
checkCriteria(criteria, unique = FALSE)

assertCriteria(
  criteria,
  unique = FALSE,
  .var.name = checkmate::vname(criteria),
  add = NULL
)

testCriteria(criteria, unique = FALSE)

checkCriterion(criterion)

assertCriterion(criterion, .var.name = checkmate::vname(criterion), add = NULL)

testCriterion(criterion)

checkCompatibleCriteria(
  criteria,
  distances = NULL,
  d = NULL,
  data = NULL,
  linkage = NULL,
  unique = FALSE
)

assertCompatibleCriteria(
  criteria,
  distances = NULL,
  d = NULL,
  data = NULL,
  linkage = NULL,
  unique = FALSE,
  .var.name = checkmate::vname(criteria),
  add = NULL
)

testCompatibleCriteria(
  criteria,
  distances = NULL,
  d = NULL,
  data = NULL,
  linkage = NULL,
  unique = FALSE
)

checkCompatibleCriterion(
  criterion,
  distances = NULL,
  d = NULL,
  data = NULL,
  linkage = NULL
)

assertCompatibleCriterion(
  criterion,
  distances = NULL,
  d = NULL,
  data = NULL,
  linkage = NULL,
  .var.name = checkmate::vname(criterion),
  add = NULL
)

testCompatibleCriterion(
  criterion,
  distances = NULL,
  d = NULL,
  data = NULL,
  linkage = NULL
)
}
\arguments{
\item{.var.name}{Name of the checked object to print in assertions.
Defaults to the heuristic implemented in \code{\link[checkmate:vname]{checkmate::vname()}}.}

\item{add}{Collection to store assertion messages.
See \link[checkmate:AssertCollection]{checkmate::AssertCollection}.}
}
\value{
Please see \code{\link[checkmate:checkMatrix]{checkmate::checkMatrix()}} help page (for instance)
for further details about the difference between check_, test_ and assert_
functions results.
}
\description{
Checking is an argument is a criterion
Check if an argument is a correct clustering criterion.
}
\details{
Extension of the \code{checkmate} package.
}
\section{Functions}{
\itemize{
\item \code{checkCriteria()}: Check if criteria are actually available

\item \code{checkCriterion()}: Check if ONE criterion is actually available

\item \code{checkCompatibleCriteria()}: Check if criteria are compatible with given data.

}}
\keyword{internal}
