% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{checkPartition}
\alias{checkPartition}
\alias{assertPartition}
\alias{testPartition}
\title{Check if an argument is a partition}
\usage{
checkPartition(x, n = length(x))

assertPartition(x, n = length(x), .var.name = checkmate::vname(x), add = NULL)

testPartition(x, n = length(x))
}
\arguments{
\item{x}{Argument to be tested}

\item{n}{the expected length of the partition.}

\item{.var.name}{Name of the checked object to print in assertions.
Defaults to the heuristic implemented in \code{\link[checkmate:vname]{checkmate::vname()}}.}

\item{add}{Collection to store assertion messages.
See \link[checkmate:AssertCollection]{checkmate::AssertCollection}.}
}
\value{
Please see \code{\link[checkmate:checkMatrix]{checkmate::checkMatrix()}} help page (for instance)
for further details about the difference between check_, test_ and assert_
functions results.
}
\description{
A partition is a Partition object or a non-empty integer-ish vector
with no missing values.
}
\details{
Extension of the \code{checkmate} package.
}
\examples{
checkPartition(c(1, 2, 3, 1))
checkPartition(matrix(1L:4L, nrow = 2L))
}
\seealso{
Other arguments checkers: 
\code{\link{checkContiguityGraph}()},
\code{\link{checkContiguityMatrix}()},
\code{\link{checkDistanceMatrix}()},
\code{\link{checkRegionalisation}()}
}
\concept{arguments checkers}
\keyword{internal}
