% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{checkRegionalisation}
\alias{checkRegionalisation}
\alias{assertRegionalisation}
\alias{testRegionalisation}
\title{Check if an argument is a regionalisation}
\usage{
checkRegionalisation(regionalisation, contiguity)

assertRegionalisation(
  regionalisation,
  contiguity,
  .var.name = checkmate::vname(regionalisation),
  add = NULL
)

testRegionalisation(regionalisation, contiguity)
}
\arguments{
\item{regionalisation}{The argument to check.}

\item{contiguity}{A contiguity matrix or a contiguity graph (\code{igraph}).}

\item{.var.name}{Name of the checked object to print in assertions.
Defaults to the heuristic implemented in \code{\link[checkmate:vname]{checkmate::vname()}}.}

\item{add}{Collection to store assertion messages.
See \link[checkmate:AssertCollection]{checkmate::AssertCollection}.}
}
\value{
Please see \code{\link[checkmate:checkMatrix]{checkmate::checkMatrix()}} help page (for instance)
for further details about the difference between check_, test_ and assert_
functions results.
}
\description{
A regionalisation if a partition with all clusters being connected
(there only have one connected component each), regarding to a contiguity
relation.
}
\details{
Extension of the \code{checkmate} package.
}
\examples{
contiguity <- simple_grid(3L, 2L)$contiguity

regionalisation <- c(1L, 1L, 1L,
                     2L, 2L, 2L)

checkRegionalisation(regionalisation, contiguity) # TRUE

notRegionalisation <- c(1L, 2L, 1L,
                        1L, 2L, 1L)

checkRegionalisation(notRegionalisation, contiguity) # message
}
\seealso{
Other arguments checkers: 
\code{\link{checkContiguityGraph}()},
\code{\link{checkContiguityMatrix}()},
\code{\link{checkDistanceMatrix}()},
\code{\link{checkPartition}()}
}
\concept{arguments checkers}
\keyword{internal}
