% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{clustering_criterion}
\alias{clustering_criterion}
\title{Evaluating quality of partitions}
\usage{
clustering_criterion(
  partitions,
  criterion,
  distances = NULL,
  d = NULL,
  data = NULL,
  standardQuant = FALSE,
  binarQual = FALSE,
  linkage = NULL,
  contiguity = NULL,
  connected = FALSE,
  ...
)
}
\arguments{
\item{partitions}{a partition or a list of partitions of the same set
the criterion will be calculated on. (vector or list of vectors)}

\item{criterion}{one of the available criterion. Use \code{\link[=available_criteria]{available_criteria()}}
to see what criterion can be applied. (string)}

\item{distances}{The distance matrix of the problem. This can be omitted
if a distance function \code{d} and data context \code{data} are provided. If only
\code{distances} is provided, all distances must be present. (distance matrix)}

\item{d}{Distance function between elements. Some criteria require this value
.If present, \code{data} must also be
specified. Some classical distances are available, it is recommended to use
them rather than a personal function for optimization reasons :
\itemize{
\item "\code{euclidean}": euclidean distance.
\item "\code{manhattan}" : manhattan distance.
\item "\code{minkowski}" : minkowski distance. In that case a value for p >= 1
must be specified.
(function or string)
}}

\item{data}{a data.frame where each row represents data related to an
element. This can be omitted if \code{d} is omitted, but might be necessary for
some criteria (e.g. Calinski-Harabasz). The present variables can
be quantitative or qualitative. If qualitative variables are present,
some distances and criteria may not be used. Possibility of standardising
variables and transforming qualitative variables into binary
variables (one-hot encoding)
using \code{standardQuant} and \code{binarQual}. (data.frame)}

\item{standardQuant}{\code{TRUE} if the variables in \code{data} should be
standardised (i.e., centered and scaled), \code{FALSE} (default) otherwise.
Standardisation is applied after the possible binarization of qualitative
variables (see \code{binarQual}). (flag)}

\item{binarQual}{\code{TRUE} if qualitative variables should be binarized (one-hot
encoding), for example, to make the data set compatible with common distances
or to standardize these variables. \code{FALSE} (default) otherwise. (flag)}

\item{linkage}{a distance linkage. Can be a string
(see \code{\link[=available_linkages]{available_linkages()}}) or a user function. Used for some of the
criteria (e.g. Dunn).}

\item{contiguity}{A contiguity matrix or an \code{igraph} contiguity graph. If not
provided, the problem is considered completely contiguous (all elements are
neighbors of each other).}

\item{connected}{a flag equals to \code{TRUE} if the criterion should be
calculated in its connected form, i.e. the mean of its value on each
connected component.}

\item{...}{Arguments specific for the criterion.}
}
\description{
This function allow to evaluate some internal clustering
criteria on one or several partitions. Connectivity can be taken into
account for reduce the impact of contiguity / connectivity constraint.
}
\seealso{
\code{\link[=available_criteria]{available_criteria()}}
}
