% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters_contiguity.R
\name{clusters_contiguity_list}
\alias{clusters_contiguity_list}
\alias{clusters_contiguity_matrix}
\title{List Contiguous Classes}
\usage{
clusters_contiguity_list(partition, x)

clusters_contiguity_matrix(partition, elemsContiguity, diag = FALSE)
}
\arguments{
\item{partition}{Partition vector. Must be numeric, containing integers,
and non-empty. (numeric vector)}

\item{x}{A contiguity matrix or a contiguity graph.
The number of elements in the matrix/graph should correspond to the size
of partition.
(matrix / contiguity graph)}
}
\value{
A 2-column matrix containing a list of contiguous classes.
By definition, a class is contiguous with itself, so this data is not stored
for memory efficiency. Similarly, only pairs (i,j) where i and j
are contiguous and i < j are stored.
}
\description{
Function returning the set of contiguous classes in a problem.
}
\section{Functions}{
\itemize{
\item \code{clusters_contiguity_matrix()}: Returns the contiguity matrix
instead of a list.

}}
\keyword{internal}
