% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHCTree.R
\name{clusters_sizes,AHCTree-method}
\alias{clusters_sizes,AHCTree-method}
\title{Clusters sizes of a AHC tree partitions.
Return in a list the sizes of the clusters of each partition
of an \code{AHCTree}.}
\usage{
\S4method{clusters_sizes}{AHCTree}(partition, sizes = rep(1, length(partition)))
}
\arguments{
\item{partition}{An \code{AHCTree}.}

\item{sizes}{A numeric vector representing the sizes of
individual elements.}
}
\value{
a list with for each partition of the \code{AHCTree}
a vector containing the size of each cluster.
}
\description{
Clusters sizes of a AHC tree partitions.
Return in a list the sizes of the clusters of each partition
of an \code{AHCTree}.
}
\keyword{internal}
