% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_class.R
\name{clusters_sizes,Partition-method}
\alias{clusters_sizes,Partition-method}
\title{Get the sizes of clusters of the \code{Partition} object.}
\usage{
\S4method{clusters_sizes}{Partition}(partition, sizes = rep(1, length(partition)))
}
\arguments{
\item{partition}{A Partition object.}

\item{sizes}{A numeric vector representing the sizes of
individual elements.}
}
\value{
A numeric vector representing the sizes of clusters.
}
\description{
Get the sizes of clusters of the \code{Partition} object.
}
\keyword{internal}
