% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_constraints.R
\name{clusters_sizes}
\alias{clusters_sizes}
\title{Cluster Sizes}
\usage{
clusters_sizes(partition, sizes = NULL)
}
\arguments{
\item{partition}{A numeric vector representing the partition of
the elements.}

\item{sizes}{A numeric vector representing the sizes of
individual elements.}
}
\description{
Determines the sizes of clusters for a given
partition and elements sizes.
}
