% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{complete_contiguity}
\alias{complete_contiguity}
\alias{complete_contiguity_matrix}
\alias{complete_contiguity_graph}
\title{Object representing complete contiguity}
\usage{
complete_contiguity_matrix(x)

complete_contiguity_graph(x)
}
\arguments{
\item{x}{argument giving information about the number of elements.
Can be an strictly positive integer, a \code{pbCon} object or a matrix.}
}
\description{
There is complete contiguity when all elements in a set are directly
connected to the others.
}
\seealso{
\code{\link[igraph:make_full_graph]{igraph::make_full_graph()}}
}
\keyword{internal}
