% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{connected_components}
\alias{connected_components}
\alias{nb_connected_components}
\title{Connected Components}
\usage{
connected_components(contiguity)

nb_connected_components(contiguity)
}
\arguments{
\item{contiguity}{The contiguity matrix of the \code{n} elements.}
}
\value{
A integer vector of length \code{n} with for each element the id of
its connected component.
}
\description{
Determines the set of connected components in a set via
its contiguity matrix.
}
\section{Functions}{
\itemize{
\item \code{nb_connected_components()}: Returns the number of
connected components in the set.

}}
\keyword{internal}
\keyword{internals}
