% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_class.R
\name{constructor_Partition}
\alias{constructor_Partition}
\title{Create a \code{Partition} using the provided parameters.}
\usage{
constructor_Partition(
  pb,
  partition,
  method,
  methodDetails = list(),
  contiguity = NA,
  minConstraint = NA_real_,
  maxConstraint = NA_real_,
  scoreMinSizeConst = NA_real_,
  scoreMaxSizeConst = NA_real_,
  criteria = list()
)
}
\arguments{
\item{pb}{A \code{pbCon} object.}

\item{partition}{A numeric vector representing the partition.}

\item{method}{A character string representing the method used
for partitioning.}

\item{methodDetails}{A list containing details about the method used
for partitioning.}

\item{contiguity}{A logical value indicating if the partition satisfies
the contiguity constraint.}

\item{minConstraint}{A logical value indicating if the partition satisfies
the min size constraint.}

\item{maxConstraint}{A logical value indicating if the partition satisfies
the max size constraint.}

\item{scoreMinSizeConst}{A numeric value representing the minimum
score for size constraints. If \code{NA} this value is calculated.}

\item{scoreMaxSizeConst}{A numeric value representing the maximum
score for size constraints. If \code{NA} this value is calculated.}

\item{criteria}{A list containing criteria for the partition.}
}
\value{
A new Partition object.
}
\description{
Create a \code{Partition} using the provided parameters.
}
\keyword{internal}
