% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{contiguity_conversion}
\alias{contiguity_conversion}
\alias{contiguity_matrix_to_graph}
\alias{graph_to_contiguity_matrix}
\title{Converts contiguity between \code{igraph} and matrix}
\usage{
contiguity_matrix_to_graph(x)

graph_to_contiguity_matrix(x)
}
\arguments{
\item{x}{object to be converted.}
}
\description{
Converts contiguity between \code{igraph} and matrix
}
\keyword{internal}
