% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_nearest_neighbor}
\alias{cpp_nearest_neighbor}
\alias{nearest_neighbor_matrix}
\title{Find the nearest neighbor of some points}
\usage{
nearest_neighbor_matrix(
  distances,
  subsetPoints,
  subsetNeighbors,
  contiguity,
  inner
)
}
\arguments{
\item{distances}{distances between points and neighbors. A row is for a
point and a column a neighbor.}

\item{subsetPoints}{A vector of strictly positive integers giving the
position of the points the nearest neighbor must be found. Positions
starts from 1 like in R. Must be at least of length one and it would be
better if values are unique.}

\item{subsetNeighbors}{A vector of strictly positive integers giving the
position of the elements of the set that can be consider has a neighbor
for each point of \code{subsetPoints}. Positions starts from 1 like in R.
Must be at least of length one and it would be
better if values are unique.}

\item{contiguity}{In the case there are contiguity constraints
between points. If an element of \code{subsetPoints} and one
of \code{subsetNeighbors} are not contiguous then the second element cannot
be a neighbor of the first one. If no contiguity constraints must be
considered a matrix with 0 rows and 0 columns must be sent. Otherwise
must be a logical matrix with same number of rows and columns than
\code{distante}.}
}
\value{
a vector of length \code{p} with for each point the position of its
nearest neighbor (0 if it doesn't exist). Positions starts from 1 like in R.
}
\description{
For a set of points (\code{subsetPoints}), find the nearest
neighbor in another set (\code{subsetNeighbors}). Distances are given in a
\verb{p' x n'} matrix (\code{distances}), with \verb{p' >= p} (p the number of points)
and \verb{n' >= n} (n the number of neighbors). If points and neighbors come from
the same set, put inner to \code{TRUE} for assuring the nearest neighbor of a
point will not be itself.
}
\section{Functions}{
\itemize{
\item \code{nearest_neighbor_matrix()}: case when \code{distances} is a matrix

}}
\keyword{internal}
