% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_elements.R
\name{diameter_set}
\alias{diameter_set}
\alias{diameter_cluster}
\title{Calculate the diameter of a set (or subset) of elements.}
\usage{
diameter_set(distances)

diameter_cluster(distances, indices)
}
\arguments{
\item{distances}{A distance matrix for the set of elements.}

\item{indices}{Indices or names of the elements that
are members of the cluster.}
}
\value{
The diameter of the set (or subset) as a positive real number.
}
\description{
Those functions calculates the diameter (i.e. the maximum distance between
two elements of the set) of a set (or subset) of elements
based on a distance matrix.
}
\section{Functions}{
\itemize{
\item \code{diameter_set()}: get the diameter of an entire set.

\item \code{diameter_cluster()}: calculates the diameter of a subset
defined by \code{indices}.

}}
