% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat.R
\name{distMat-access}
\alias{distMat-access}
\alias{[,DistMat,matrix,missing,ANY-method}
\alias{[,DistMat,numericOrMissing,numericOrMissing,ANY-method}
\alias{[,DistMat,list,missing,ANY-method}
\title{Access to \code{DistMat} data`}
\usage{
\S4method{[}{DistMat,matrix,missing,ANY}(x, i, j, drop)

\S4method{[}{DistMat,numericOrMissing,numericOrMissing,ANY}(x, i, j, drop)

\S4method{[}{DistMat,list,missing,ANY}(x, i, j, drop)
}
\arguments{
\item{x}{An \code{AbstractSymMat} or inheriting object}

\item{i}{A vector composed of the indices of the wished rows.}

\item{j}{A vector composed of the indices of the wished columns}
}
\value{
Depending on \code{drop} and the dimensions of \code{i} and \code{j}, a matrix or
a vector.
}
\description{
Rules are the following :
\itemize{
\item when rows and columns are given a submatrix is returned with the
asked rows and columns
\item when only rows (resp. columns are given) a submatrix is return with
the asked rows (resp. columns) and the entire columns (resp. rows)
\item When rows and columns are missing the entire matrix is sent back.
Depending on \code{drop} (a flag) is there is only one row and / or
one column given the result will be simplified in a vector form or not.
If some values are missing they will be calculated if possible with the \code{d}
distance function with the \code{data} data frame.
}
}
\keyword{internal}
