% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat.R
\name{distMat_access_diag}
\alias{distMat_access_diag}
\title{Access to diagonal of a \code{DistMat} object. The result is constantly
a double vector of length the number of elements.}
\usage{
\S4method{diag}{DistMat}(x)
}
\description{
Access to diagonal of a \code{DistMat} object. The result is constantly
a double vector of length the number of elements.
}
\keyword{internal}
