% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat.R
\name{distMat_alleq}
\alias{distMat_alleq}
\alias{all.equal,matrix,DistMat-method}
\alias{all.equal,DistMat,DistMat-method}
\title{Custom all.equal method for comparing \code{DistMat} objects with matrices}
\usage{
\S4method{all.equal}{DistMat,matrix}(target, current, ...)

\S4method{all.equal}{matrix,DistMat}(target, current, ...)

\S4method{all.equal}{DistMat,DistMat}(target, current, ...)
}
\arguments{
\item{target, current}{two variables, one of them at least being a \code{DistMat}.}
}
\value{
TRUE if the target and current are equal, FALSE otherwise
}
\description{
Custom all.equal method for comparing \code{DistMat} objects with matrices
}
\section{Functions}{
\itemize{
\item \code{all.equal(target = matrix, current = DistMat)}: matrix + \code{DistMat}

\item \code{all.equal(target = DistMat, current = DistMat)}: \code{DistMat} + \code{DistMat}

}}
\keyword{internal}
