% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat.R
\name{distMat_NA}
\alias{distMat_NA}
\alias{anyNA,DistMat-method}
\alias{is.na,DistMat-method}
\title{Is there missing values in the distance matrix ?}
\usage{
\S4method{anyNA}{DistMat}(x)

\S4method{is.na}{DistMat}(x)
}
\arguments{
\item{x}{a \code{DistMat} object}
}
\description{
Is there missing values in the distance matrix ?
}
\section{Functions}{
\itemize{
\item \code{anyNA(DistMat)}: Check if the \code{DistMat} object contains any NA values

\item \code{is.na(DistMat)}: Check what values are NA

}}
\keyword{internal}
