% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunn_index.R
\name{dunn}
\alias{dunn}
\alias{dunn_index}
\title{Dunn Index}
\usage{
dunn_index(distances, partition, linkage = "single", valueOnly = FALSE, ...)
}
\arguments{
\item{distances}{The inter-element distance matrix.}

\item{partition}{The partition of interest.}

\item{linkage}{The method for computing inter-cluster distances.
Defaults to "single" (minimum distance between clusters).}

\item{valueOnly}{If TRUE, returns only the Dunn index value (D)
instead of the full result.}
}
\value{
If \code{valueOnly} is TRUE, returns the Dunn index value (D) as a
floating-point number. Otherwise, returns a list containing the
following components:
- \code{diameters}: A vector containing the diameters of each cluster.
- \code{linkageDistances}: A matrix containing the inter-cluster
distances.
- \code{clustersMoinsSepares}: The IDs of the two clusters with
the smallest separation.
- \code{separation}: The separation between the two clusters with
the smallest separation.
- \code{leastCompactCluster}: The ID of the least compact cluster.
- \code{compactness}: The diameter of the least compact cluster.
- \code{D}: The Dunn index value, calculated as the separation divided
by the compactness.
- \code{linkage}: The method used for computing inter-cluster distances.
}
\description{
Calculates the Dunn index for a given partition based on
the inter-element distance matrix. The Dunn index assesses the compactness
and separation of clusters in a partition.
}
\references{
Dunn, Joseph C. (1973) A fuzzy relative of the ISODATA
process and its use in detecting compact well-separated clusters,
Journal of Cybernetics
}
\keyword{internal}
