% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transferable_points.R
\name{elements_transferables}
\alias{elements_transferables}
\alias{transferable_elements_pb}
\alias{maj_elements_transferables}
\title{Generates a list of transferable elements for \code{pbCon} class}
\usage{
transferable_elements_pb(regionalisation, pb, contiguityGraph, contiguityMat)

maj_elements_transferables(
  anciennePartition_vec,
  nouvellePartition_vec,
  oldTransferablePoints_list,
  pb,
  contiguityGraph,
  contiguityMat,
  donor = NULL,
  receiver = NULL
)
}
\arguments{
\item{regionalisation}{The partition (vector of positive integers)}

\item{pb}{A \code{pbCon} class object}

\item{contiguityGraph}{An \code{igraph} contiguity graph}

\item{contiguityMat}{A \code{contiguity} matrix}

\item{anciennePartition_vec}{The initial partition
(vector of positive integers)}

\item{nouvellePartition_vec}{The modified partition
(vector of positive integers)}

\item{oldTransferablePoints_list}{The previous list of
transferable elements}

\item{donor}{Identifier of the donor cluster}

\item{receiver}{Identifier of the receiver cluster}
}
\value{
A list with three elements: 'frontier', 'articulation',
'transferable' containing information about boundary, articulation,
and transferable elements.
}
\description{
Determines the list of transferable elements for each cluster in a partition.

Determines the updated list of transferable elements after a modification
in the partition.
}
\section{Functions}{
\itemize{
\item \code{maj_elements_transferables()}: Determines the updated list of
transferable elements after a modification in the partition.

}}
\keyword{internal}
