% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_tools.R
\name{equivalent_partitions}
\alias{equivalent_partitions}
\title{Are two partitions equivalent?}
\usage{
equivalent_partitions(partition1, partition2)
}
\arguments{
\item{partition1, partition2}{two non-empty vectors representing.
partitions to be compared}
}
\value{
\code{TRUE} if both partitions are equivalent, \code{FALSE} otherwise.
}
\description{
Check if two partitions are equivalent, i.e. represent the same clustering.
}
