% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_grid.R
\name{gen_grid}
\alias{gen_grid}
\alias{simple_grid}
\title{Grid Simulation}
\usage{
gen_grid(
  x_int,
  y_int,
  contiguityType = "Queen",
  nbIndividuals = x_int * y_int,
  nbMinEmptyZones = 0L,
  nbMetropolises = 0L,
  propInMetropolises = 0.4,
  quantitatives_mat = c(0, 1),
  qualitatives_list = NULL,
  nbQuantitatives = 0L,
  seed = NULL
)

simple_grid(x_int, y_int = x_int, avgPersonsPerCell = 100, seed = NULL)
}
\arguments{
\item{x_int, y_int}{dimensions of the grid (positive integers)}

\item{contiguityType}{type of contiguity to apply for the grid.
Two options:
\itemize{
\item "Rook": two cells are contiguous if they share a side.
\item "Queen": two cells are contiguous if they share a side or a vertex.
}}

\item{nbIndividuals}{Number of individuals to distribute
(strictly positive integer).}

\item{nbMinEmptyZones}{Number of zones where no individuals should
be present. Default is 0. \code{nbMinEmptyZones + nbMetropolises} must be smaller
than \code{x_int * y_int}.
(positive integer)}

\item{nbMetropolises}{Number of zones that should be more populated.
Default is 0. \code{nbMinEmptyZones_int + nbMetropolises}_int must be smaller than
\code{x_int * y_int}. (positive integer)}

\item{propInMetropolises}{Proportion of individuals found in metropolises
(if any). (floating point in \verb{[0,1]})}

\item{quantitatives_mat}{Characteristics of quantitative variables
to generate.
If it's a vector, it must be of size 2. The first element indicates the mean,
and the second element indicates the standard deviation.
If \code{nbQuantitatives > 0},
as many i.i.d. variables will be generated with the data specified
in the vector.
If it's a matrix, it should have 2 columns. The first column corresponds
to the means of the independent quantitative variables to be generated.
The second column corresponds to the standard deviation of these variables.
Each row corresponds to a variable.
All quantitative variables follow a normal distribution.
(vector of size 2, matrix with 2 columns, or NULL)}

\item{qualitatives_list}{List of vectors where each vector contains
the categories of the qualitative variables to generate. Each qualitative
variable will be generated by drawing uniformly from its categories.
(list of vectors or NULL)}

\item{nbQuantitatives}{If \code{quantitatives_mat} is a vector, indicates
the number of quantitative variables to generate with the characteristics
defined in \code{quantitatives_mat}. Ignored otherwise. (positive or zero integer)}

\item{seed}{indicates if the random seed should be fixed before generating
values. \code{NULL} (default) if the seed must not be fixed. Otherwise must
be an integer. (integer)}
}
\value{
a list including:
\enumerate{
\item \code{contiguity}: grid's contiguity matrix
\item \code{distribution}: a dataframe specifying the distribution of sizes
across different cells. Contains:
\itemize{
\item \code{nbIndividuals}: number of individuals in the zone. The sum equals
\code{nbElems_int}.
\item \code{emptyFixedZone}: \code{TRUE} if the zone has been fixed as empty,
\code{FALSE} otherwise.
\item \code{metropolis}: \code{TRUE} if the zone is a metropolis, \code{FALSE} otherwise.
}
\item \code{context}: a dataframe where each row corresponds to the context/data
related to an individual, and each column is a variable.
}
}
\description{
Simulates a territory in the form of a grid with its contiguity,
a distribution of individuals across the cells, and a context for
different sizes.
}
\section{Functions}{
\itemize{
\item \code{simple_grid()}: simpler version with "Queen" contiguity,

}}
\keyword{internal}
