% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation.R
\name{gen_initial_partition}
\alias{gen_initial_partition}
\alias{gen_initial_partitions}
\title{Generate initial partition.}
\usage{
gen_initial_partition(pb, mode = "unitary", nbFusions = 1L)

gen_initial_partitions(pb, modes, nbFusions = 1L)
}
\arguments{
\item{pb}{A \code{pbCon} object.}

\item{mode}{A character string specifying the mode of generating
the initial partition. Possible values are
"unitary" (default) or "random".}

\item{nbFusions}{An integer specifying the number of fusions
to be performed in "random" mode.}

\item{modes}{A character vector specifying the modes of generating
the initial partitions.}
}
\value{
A numeric vector representing the initial partition.
}
\description{
This function generates an initial partition considering
connectivity and maximum size constraints.
}
\section{Functions}{
\itemize{
\item \code{gen_initial_partitions()}: generates multiple
initial partitions

}}
\keyword{internal}
