% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_grid.R
\docType{data}
\name{grids}
\alias{grids}
\alias{grid_queen_vide0_metropole0_x7_y7_indivMoy100_quant3_qual0}
\alias{grid_queen_vide0_metropole0_x20_y20_indivMoy100_quant3_qual0}
\alias{grid_queen_vide0_metropole0_x30_y50_indivMoy100_quant3_qual0}
\alias{grid_queen_vide0_metropole0_x50_y70_indivMoy100_quant3_qual0}
\title{A few grids to apply connectivity constraints.}
\format{
for each grid a list with the following components:
\enumerate{
\item \code{contiguity}: grid's contiguity matrix
\item \code{distribution}: a dataframe specifying the distribution of sizes
across different cells. Contains:
\itemize{
\item \code{nbIndividuals}: number of individuals in the zone. The sum equals
\code{nbElems_int}.
\item \code{emptyFixedZone}: \code{TRUE} if the zone has been fixed as empty,
\code{FALSE} otherwise.
\item \code{metropolis}: \code{TRUE} if the zone is a metropolis, \code{FALSE} otherwise.
}
\item \code{context}: a dataframe where each row corresponds to the context/data
related to an individual, and each column is a variable.
}

An object of class \code{list} of length 3.

An object of class \code{list} of length 3.

An object of class \code{list} of length 3.

An object of class \code{list} of length 3.
}
\usage{
grid_queen_vide0_metropole0_x7_y7_indivMoy100_quant3_qual0

grid_queen_vide0_metropole0_x20_y20_indivMoy100_quant3_qual0

grid_queen_vide0_metropole0_x30_y50_indivMoy100_quant3_qual0

grid_queen_vide0_metropole0_x50_y70_indivMoy100_quant3_qual0
}
\description{
Each grid has a number of rows and a number of columns. An average of
100 people per case are dispatched around the grid. No case are fixed as
empty or a a metropolis. The context is made of 2 quantitative variables and
one qualitative.
}
\keyword{data}
\keyword{internal}
