% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interior_boundary.R
\name{interior_boundary}
\alias{interior_boundary}
\title{Interior boundary of a partition}
\usage{
interior_boundary(contiguity, partition, removeSymmetry = TRUE)
}
\arguments{
\item{contiguity}{A contiguity matrix or a contiguity graph
(package \code{igraph}). Prefer matrix is possible.
(contiguity matrix or \code{igraph} graph)}

\item{partition}{The partition. (vector of strictly positive integers)}

\item{removeSymmetry}{\code{TRUE} (default) if symmetry should be removed
from the result. (flag)}
}
\value{
a matrix with 4 columns (\code{x1}, \code{x2}, \code{cluster1}, \code{cluster2}) with
as many rows as possible connections between an element of a cluster and
another. If \code{removeSymmetry = TRUE}, only rows where \code{x1 < x2} are kept.
}
\description{
Determines the set of elements, distributed in clusters, that are
contiguous to an element from another cluster.
}
\details{
The interior boundary of a partition, under a certain
contiguity, is defined as the set of elements contiguous to at least
one other element from a different cluster than its own.
}
