% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractSymMat.R
\name{is.na,AbstractSymMat-method}
\alias{is.na,AbstractSymMat-method}
\title{Where are located NA is our symmetric matrix?}
\usage{
\S4method{is.na}{AbstractSymMat}(x)
}
\arguments{
\item{x}{An \code{AbstractSymMat} object.}
}
\value{
A logical matrix of the same dimensions as the \code{AbstractSymMat}
object, indicating if each element is NA.
}
\description{
Where are located NA is our symmetric matrix?
}
\keyword{internal}
