% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{is_connected,matrix-method}
\alias{is_connected,matrix-method}
\title{Check Connectivity}
\usage{
\S4method{is_connected}{matrix}(graph, mode = c("weak", "strong"))
}
\arguments{
\item{graph}{Contiguity matrix of the elements.
(contiguity matrix)}
}
\value{
TRUE if the set is connected, FALSE otherwise.
}
\description{
Verifies via the contiguity matrix whether a set is connected.
}
\examples{
contiguity1 <- matrix(c(FALSE, TRUE, FALSE,
                        TRUE, FALSE, TRUE,
                        FALSE, TRUE, FALSE), nrow = 3L)

is_connected(contiguity1) # TRUE

contiguity2 <- matrix(c(FALSE, FALSE,
                        FALSE, FALSE), nrow = 2L)

is_connected(contiguity2) # FALSE
}
\seealso{
\code{\link[igraph:components]{igraph::is_connected()}}
}
\keyword{internal}
