% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_linkage.R
\name{linkage_distances}
\alias{linkage_distances}
\alias{single_linkage}
\alias{complete_linkage}
\alias{centroid_linkage}
\alias{is_single_linkage}
\alias{is_complete_linkage}
\alias{is_single_or_complete_linkage}
\alias{is_average_linkage}
\alias{is_medoid_linkage}
\alias{is_centroid_linkage}
\alias{is_hausdorff_linkage}
\title{Distances between clusters (linkage distances)}
\usage{
single_linkage(distances)

complete_linkage(distances)

centroid_linkage(dataCluster1, dataCluster2, d = euclidean_distance)

is_single_linkage(...)

is_complete_linkage(...)

is_single_or_complete_linkage(linkage)

is_average_linkage(...)

is_medoid_linkage(...)

is_centroid_linkage(...)

is_hausdorff_linkage(...)
}
\arguments{
\item{distances}{the distance matrix between elements of the first and the
second cluster. (matrix)}

\item{dataCluster1}{Data related to elements of the first cluster.
Must contain at least one individual and at least one variable. All variables
must be quantitative. (data frame)}

\item{dataCluster2}{Data related to elements of the second cluster.
The restrictions are the same as for \code{dataCluster1}. Must contain the same
variables as the first data frame. (data frame)}

\item{d}{The distance function between elements. Apart from the first two
arguments, all others (if any) must be optional. Must return a positive real
number. Default is the euclidean distance. (function)}
}
\value{
The distance between the two clusters. (one positive value)
}
\description{
Some functions for calculating distances between two sets
using the distances between elements.
}
\section{Functions}{
\itemize{
\item \code{single_linkage()}: do the single linkage, i.e. take the minimum
of the distances between the elements of the first and the elements of the
second cluster.

\item \code{complete_linkage()}: do the complete linkage, i.e. take the maximum
of the distances between the elements of the first and the elements of the
second cluster.

\item \code{centroid_linkage()}: do the centroid linkage distance. Distance
used to obtain a distance on classes from elements composed of
quantitative variables. The centroid of each cluster is determined, and the
inter-element distance is applied between these two centroids.

\item \code{is_single_linkage()}: Check if the given linkage distance is
the single linkage

\item \code{is_complete_linkage()}: Check if the given linkage distance is
the complete linkage

\item \code{is_single_or_complete_linkage()}: Check if the given linkage distance is
the single or the complete linkage

\item \code{is_average_linkage()}: Check if the given linkage distance is
the average linkage

\item \code{is_medoid_linkage()}: Check if the given linkage distance is
the medoid linkage

\item \code{is_centroid_linkage()}: Check if the given linkage distance is
the centroid linkage

\item \code{is_hausdorff_linkage()}: Check if the given linkage distance is
the Hausdorff linkage

}}
\keyword{internal}
