% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medoids.R
\name{medoid}
\alias{medoid}
\alias{medoids_partition}
\title{Medoid Calculation}
\usage{
medoid(distances, byName = FALSE)

medoids_partition(distances, partition, clusters = NULL)
}
\arguments{
\item{distances}{The distance matrix used to find the
medoid(s) (distance matrix).}

\item{byName}{\code{TRUE} if the returned result should be the row name
corresponding to the medoid, \code{FALSE} if it should be the row number.
Ignored if \code{distances} is not named. (flag)}

\item{partition}{The partition for which the medoids of each cluster
should be determined (vector).}

\item{clusters}{The set of clusters for which the medoid should be
calculated (all by default).}
}
\value{
For \code{medoid}: depending on \code{byName}, the name or the
identifier of the medoid.

For \code{medoids_partition}: a named numeric vector indicating,
for each cluster or a subset of them, the identifier of its medoid.
}
\description{
Determines the medoid of one or several subsets of
elements using a distance matrix.
}
\section{Functions}{
\itemize{
\item \code{medoid()}: Determine the medoid of a set.

\item \code{medoids_partition()}: Calculation of the medoid for each cluster
in a partition.

}}
\examples{
M <- matrix(c(0.0, 1.0, 2.0,
              1.0, 0.0, 3.0,
              2.0, 3.0, 00.))

medoid(M)
}
