% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_tools.R
\name{merge_cc_partitions}
\alias{merge_cc_partitions}
\title{Merges Partition Connected Components}
\usage{
merge_cc_partitions(components, partitions)
}
\arguments{
\item{components}{A vector indicating the connected component index of
each element.}

\item{partitions}{A list of partition vectors, where each vector represents
a partition of a connected component.}
}
\value{
A merged partition vector.
}
\description{
This function merges the partitions of connected components back together.
}
