% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_elements.R
\name{minkowski_distance}
\alias{minkowski_distance}
\title{Calculate the distance between two elements using the Minkowski distance.}
\usage{
minkowski_distance(x, y, p = 2L)
}
\arguments{
\item{x}{First element (numeric vector of non-zero length).}

\item{y}{Second element. Must be of the same length as x
(numeric vector of non-zero length).}

\item{p}{Real number greater than or equal to one.}
}
\value{
A positive real number representing the Minkowski
distance between x and y.
}
\description{
Calculate the distance between two elements using the Minkowski distance.
}
\keyword{internal}
