% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_tools.R
\name{nbClusters}
\alias{nbClusters}
\title{Number of clusters in a partition (per connected component)}
\usage{
nbClusters(partition, components = NULL)
}
\arguments{
\item{partition}{a partition vector.}

\item{components}{a vector giving for each element its connected component.
If \code{NULL} (default) components will be ignored. Otherwise must be a vector
of the same length of \code{partition}.}
}
\value{
If \code{components} is \code{NULL}, the total number of clusters in
\code{partition} (positive integer). Otherwise a named vector of positive integers
of length the number of connected components and giving the number of
clusters per connected component.
}
\description{
Give the number of clusters in a partition. This number can be the
total number or the number in each connected component.
}
\examples{
nbClusters(c(1L, 2L, "x", 1L)) # 3
}
