% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity.R
\name{nearest_neighbor}
\alias{nearest_neighbor}
\title{Find the nearest neighbor of some points}
\usage{
nearest_neighbor(
  distances,
  inner = FALSE,
  contiguity = NULL,
  subsetPoints = seq_len(nrow(distances)),
  subsetNeighbors = seq_len(ncol(distances))
)
}
\arguments{
\item{distances}{distances between points and neighbors. A row is for a
point and a column a neighbor. If a vector of length \code{k > 0} is given,
it's consider like a matrix of distances with \verb{p' = k} points and \verb{n' = 1}
neighbor.}

\item{inner}{Flag indicating if points and neighbors come from the same
set and \code{distance} is the set matrix distance. Diagonal of \code{distance}
will then not be considered, assuring that the nearest neighbor of a point is
not itself. \code{FALSE} by default (different sets).}

\item{contiguity}{In the case there are connectivity constraint
between points. If an element of \code{subsetPoints} and one of
\code{subsetNeighbors} are not contiguous then the second element cannot be a
neighbor of the first one. \code{NULL} by default (no connectivity constraint).
If precised, must be a logical matrix with the same dimensions
as \code{distances}.}

\item{subsetPoints}{A vector of strictly positive integers giving the
position of the points the nearest neighbor must be found. Default is
all points (\verb{p = p'}). If \code{distances} has row names can be those instead
(as characters). If no points are given, return an empty vector.
Values can be duplicated but it will increase complexity.}

\item{subsetNeighbors}{A vector of strictly positive integers giving the
position of the elements of the set that can be consider has a neighbor
for each point of \code{subsetPoints}. Default is all neighbors (\verb{n = n'}).
If \code{distances} has column names can be those instead (as characters).
If no neighbors are given an error is generated. Values are expected
to be unique.}
}
\value{
a vector of length \code{p} with for each point the position of its
nearest neighbor (\code{NaN} if it doesn't exist). A point do not have a nearest
neighbor if each distance with its neighbors are missing. If \code{distances} has
column names position of neighbors will be replaced by their names. The
returned vector is named with the IDs of the points, or their name if
\code{distances} has row names.
}
\description{
For a set of points (\code{subsetPoints}), find the nearest
neighbor in another set (\code{subsetNeighbors}). Distances are given in a
\verb{p' x n'} matrix (\code{distances}), with \verb{p' >= p} (p the number of points)
and \verb{n' >= n} (n the number of neighbors). If points and neighbors come from
the same set, put inner to \code{TRUE} for assuring the nearest neighbor of a
point will not be itself.
}
