% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{normalize_df}
\alias{normalize_df}
\alias{numeric_variabes}
\alias{all_numeric_variables}
\alias{is_df_normalized}
\title{Normalization of data}
\usage{
normalize_df(df, standardQuant = TRUE, binarQual = FALSE)

numeric_variabes(df)

all_numeric_variables(df)

is_df_normalized(df)
}
\arguments{
\item{df}{A data frame of equivalent (\link[tibble:tibble-package]{tibble},
data table...)}

\item{standardQuant}{\code{TRUE} (default) if quantitative variables of \code{df} must
be normalised / standardised, i.e centered (having a 0 empirical mean)
and reduced / scaled (having a 1 unbiased empirical standard deviation).
\code{FALSE} otherwise. If the number of rows is equal to 1 standardisation isn't
realized. Standardisation is realized after transformation
from qualitative to quantitative, so if \code{binarQual} is equal to \code{TRUE}
transformed qualitative variables will be normalized.}

\item{binarQual}{\code{TRUE} if non-numeric values must be converted in numeric
form by creating for each qualitative variable one dummy variable
per modality minus one. Original columns will be removed.
\code{FALSE} (default) otherwise. Ignore if all variables
are quantitative. If \code{standardQuant} equals to \code{TRUE} created dummies
will be standardised as well.}
}
\value{
For \code{normalize_df} the normalized version of \code{df}. If
\code{binarQual = TRUE} and some categorical variables were if \code{df} the number
of variables (columns) of \code{df} might be higher.

For \code{numeric_variables} a logical vector of length \code{ncol(df)}
indicating which variable is numeric and which is not.

For \code{all_numeric_variables}, \code{TRUE} if all variables of \code{df}
are quantitative, \code{FALSE} otherwise. Returns \code{TRUE} by convention if there
is no variable at all.

For \code{is_df_normalized}, a flag equals to \code{TRUE} if all numeric
variables are normalized, \code{FALSE} otherwise. Returns \code{TRUE} if there is
no numeric variables.
}
\description{
Normalization of data
}
\section{Functions}{
\itemize{
\item \code{normalize_df()}: Normalize quantitative and / or qualitative \code{df}
variables.

\item \code{numeric_variabes()}: Give a logical mask indicating which columns
of \code{df} are numeric.

\item \code{all_numeric_variables()}: Give a flag indicating if all variables of \code{df}
are quantitative. Shortcut for \code{all(numeric_variables(df))}.

\item \code{is_df_normalized()}: Check if the numeric variables of \code{df}
are normalized.

}}
\references{
Kaplan, J. & Schlegel, B. (2023).
\href{https://github.com/jacobkap/fastDummies}{fastDummies}: Fast Creation
of Dummy (Binary) Columns and Rows from Categorical Variables.
}
\seealso{
\code{\link[base:scale]{base::scale()}}

\code{\link[fastDummies:dummy_cols]{fastDummies::dummy_cols()}}
}
