% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{optimal_partitions}
\alias{optimal_partitions}
\title{Sort partitions by a clustering criterion}
\usage{
optimal_partitions(
  criterionValues,
  criterion,
  byName = TRUE,
  names = NULL,
  keepNA = TRUE
)
}
\arguments{
\item{criterionValues}{Vector of numerical data indicating
the \code{criterion} values for different partitions. (numeric vector)}

\item{criterion}{Name of the clustering criterion. (character string)}

\item{byName}{\code{TRUE} (default) if what should be returned is the names of
partitions rather than their identifier. \code{FALSE} otherwise. The partition
names can be specified via the \code{names} variable. If it's \code{NULL}, then the
names of the \code{criterionValues} vector will be used.
If none are available, \code{byName} is set to \code{FALSE}. (boolean)}

\item{names}{Names of the partitions. Used when \code{byName} is \code{TRUE},
ignored otherwise. If used, it must be the same size as
\code{criterionValues}. (vector or \code{NULL})}

\item{keepNA}{flag indicating if NA values should be kept in the result.
If \code{TRUE}, they are kept at the end of the vector. If \code{FALSE}, they are
removed.}
}
\value{
A vector of the same size as \code{criterionValues}, indicating
the order (from best to worst) of partitions relative to the
requested criterion. The indicated data are the identifiers of partitions
if \code{byName = FALSE}, their names otherwise. (vector)
}
\description{
Returns the order of the best partitions according to a certain criterion.
}
