% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_class.R
\name{partition_length}
\alias{partition_length}
\alias{length,Partition-method}
\title{Number of elements in a partition}
\usage{
\S4method{length}{Partition}(x)
}
\arguments{
\item{x}{A \code{Partition} object}
}
\value{
the number of elements in the partition. (positive integer)
}
\description{
Give the number of elements in a partition
}
\keyword{internal}
