% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_class.R
\name{Partition_list_to_tibble}
\alias{Partition_list_to_tibble}
\title{Convert a list of partitions to a \link[tibble:tibble-package]{tibble}}
\usage{
Partition_list_to_tibble(partitions)
}
\arguments{
\item{partitions}{A list of \code{Partition} objects.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing information
about each partition.
}
\description{
Convert a list of partitions to a \link[tibble:tibble-package]{tibble}
}
\keyword{internal}
