% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{partition_to_list}
\alias{partition_to_list}
\title{Transforms a vector of partitions into a list where each element
is a vector indicating which element of the set is present in each cluster.}
\usage{
partition_to_list(partition)
}
\arguments{
\item{partition}{A numeric vector representing the partition.}
}
\value{
A list of vectors, each vector representing a cluster with
element indices.
}
\description{
Transforms a vector of partitions into a list where each element
is a vector indicating which element of the set is present in each cluster.
}
\keyword{internal}
