% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_score.R
\name{score_size_constraints}
\alias{score_size_constraints}
\alias{score_constraints_min}
\alias{.score_constraints_min}
\alias{score_constraints_max}
\alias{.score_constraints_max}
\alias{score_constraints_table}
\alias{score_constraints}
\title{Give a score of constraint size respect}
\usage{
score_constraints_min(
  clustersSizes = clusters_sizes(partition, sizes),
  m = 0,
  sizes = NULL,
  partition = NULL,
  f = identity
)

.score_constraints_min(m, clustersSizes, f = identity)

score_constraints_max(
  clustersSizes = clusters_sizes(partition, sizes),
  M = Inf,
  partition = NULL,
  f = identity
)

.score_constraints_max(M, clustersSizes, f = identity)

score_constraints_table(scoreContraintesMin, scoreContraintesMax)

score_constraints(
  clustersSizes = clusters_sizes(partition, sizes),
  m = 0,
  M = Inf,
  partition = NULL,
  f = identity,
  details = FALSE
)
}
\arguments{
\item{clustersSizes}{the size of the different clusters. If equals to \code{NULL}
(default) sizes will be calculated.}

\item{partition}{a partition for the problem. Used only is \code{clustersSizes}
is \code{NULL}.}

\item{f}{a function that will take a vector of difference between cluster
sizes and constraint value. Must be vectorial, give 0 when values are
negative and positive otherwise.}

\item{details}{flag indicating if details must be given or only
the final score.}

\item{pb}{A \code{pbCon} object.}
}
\value{
a positive double value indicating how much the partition
respect the size constraint(s). If it is equals to 0 then all constraint
are respected by each cluster. The higher the score is the less the
constraint(s) are respected. (positive double value)
}
\description{
Those functions give a score that indicates if a partition respects
or not some size constraint(s). A score equals to 0 means that the constraint
is completely respected. A strictly positive value means that some clusters
do not respect the constraint size.
}
\keyword{internal}
