% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_constraints.R
\name{simp_cont}
\alias{simp_cont}
\alias{simplify_minSizeConst}
\alias{simplify_maxSizeConst}
\title{Simplify Minimum Size Constraint}
\usage{
simplify_minSizeConst(m_vec, sizes = NULL)

simplify_maxSizeConst(M_vec, sizes, connectedComponents = NULL)
}
\arguments{
\item{m_vec}{A numeric vector with min constraints to be simplified.}

\item{sizes}{A numeric vector representing the sizes
of elements.}

\item{M_vec}{A numeric vector with max constraints values to be simplified.}

\item{connectedComponents}{An optional numeric vector specifying
the connected components.}
}
\description{
Simplifies the minimum size constraint of a given vector.
}
\section{Functions}{
\itemize{
\item \code{simplify_maxSizeConst()}: Simplifies the maximum size constraint
of a given vector.

}}
\keyword{internal}
