% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_tools.R
\name{standardize_partition}
\alias{standardize_partition}
\title{Standardize a partition}
\usage{
standardize_partition(partition)
}
\arguments{
\item{partition}{A non-empty vector representing the partition.}
}
\value{
A strictly positive integer vector representing the updated
partition with consecutive cluster identifiers.
}
\description{
Changes the cluster identifiers in the partition vector to form a
consecutive sequence.
}
\details{
This form is also used by \code{\link[stats:cutree]{stats::cutree()}}.

Two partitions are equivalent if and only if their
standardised forms are equal.
}
