% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_grid.R
\docType{data}
\name{survey}
\alias{survey}
\alias{survey_grid_50x70}
\title{Simulation of a survey}
\format{
A list with the following components:
\enumerate{
\item \code{contiguity}: Contiguity matrix of the grid
\item \code{repartition}: A dataframe specifying the distribution of sizes
across different cells. Contains:
\itemize{
\item \code{nbIndividuals}: number of individuals in the zone. The sum is
\code{nbElems_int}.
\item \code{emptyFixedZone}: \code{TRUE} if the zone has been fixed as empty,
\code{FALSE} otherwise.
\item \code{metropolis}: \code{TRUE} if the zone is a metropolis, \code{FALSE} otherwise.
}
\item \code{context}: A dataframe where each row corresponds to the context /
data related to an individual and each column is a variable.
}
}
\usage{
survey_grid_50x70
}
\description{
This grid simulates a survey. It has a dimension of 50 * 70 (3500 cells).
Most of the cells are empty (around 80\%). The total number of people is
4000. 1\% of the cells are metropolises, aggregating 40\% of the population.
The context consists of 2 quantitative and 1 qualitative variable.
}
\keyword{data}
\keyword{internal}
