% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calinski_harabasz.R
\name{update_calinski_harabasz}
\alias{update_calinski_harabasz}
\title{Update the Calinski-Harabasz index}
\usage{
update_calinski_harabasz(
  dataCriterion,
  donor,
  receiver,
  givenElements,
  partitionBefore,
  dataElements
)
}
\arguments{
\item{dataCriterion}{a list containing all data about
the actual criterion value.}

\item{donor}{the id of the donor cluster.}

\item{givenElements}{set of the elements of \code{donor} given to
\code{receveir}. (vector of length at least 1 made of strictly positive integers)}

\item{partitionBefore}{the partition before the exchange of elements.}

\item{dataElements}{data about the elements. Corresponds to \code{data} in
other \code{c3t} functions.}
}
\description{
Update the Calinski-Harabasz index
}
\keyword{internal}
