% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{update_criterion_params}
\alias{update_criterion_params}
\title{Update criterion parameters}
\arguments{
\item{partitionBefore}{the partition before the exchange of elements.}

\item{dataCriterion}{a list containing all data about
the actual criterion value.}

\item{donor}{the id of the donor cluster.}

\item{receveir}{the id of the receiver cluster.}

\item{givenElement}{element of \code{donor} given to
\code{receveir}. (strictly positive integer)}

\item{givenElements}{set of the elements of \code{donor} given to
\code{receveir}. (vector of length at least 1 made of strictly positive integers)}

\item{dataElements}{data about the elements. Corresponds to \code{data} in
other \code{c3t} functions.}
}
\description{
Update criterion parameters
}
\keyword{internal}
