% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunn_index.R
\name{update_dunn_index}
\alias{update_dunn_index}
\title{Update Dunn index}
\usage{
update_dunn_index(
  dataCriterion,
  distances,
  partitionBefore,
  donor,
  receiver,
  givenElement
)
}
\arguments{
\item{dataCriterion}{a list containing all data about
the actual criterion value.}

\item{partitionBefore}{the partition before the exchange of elements.}

\item{donor}{the id of the donor cluster.}

\item{givenElement}{element of \code{donor} given to
\code{receveir}. (strictly positive integer)}
}
\description{
Update Dunn index
}
\keyword{internal}
