% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_constraints.R
\name{valid_cont}
\alias{valid_cont}
\alias{checkMinSizeConstraints}
\alias{checkMaxSizeConstraints}
\title{Validate Size Constraints}
\usage{
checkMinSizeConstraints(
  m_vec,
  sizes,
  connectedComponents = NULL,
  authorizeNegative = FALSE,
  len = NULL
)

checkMaxSizeConstraints(M_vec, sizes, len = NULL)
}
\arguments{
\item{m_vec}{A numeric vector with min constraints values to be tested.}

\item{sizes}{A numeric vector representing the sizes
of elements.}

\item{connectedComponents}{An optional numeric vector specifying
the connected components.}

\item{M_vec}{A numeric vector with max constraints to be tested.}
}
\description{
Validate Size Constraints
}
\section{Functions}{
\itemize{
\item \code{checkMinSizeConstraints()}: Verifies the minimum size constraint for
a given vector and its connected components.

\item \code{checkMaxSizeConstraints()}: Verifies the maximum size constraint
for a given vector.

}}
\keyword{internal}
