% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calinski_harabasz.R
\name{variance_tools}
\alias{variance_tools}
\alias{colVars}
\alias{sum_squares}
\title{Variance and sum of square
Tools for calculating variance and sum of squares
with a matrix or a data.frame.}
\usage{
colVars(data, unbiased = TRUE, aggregate = FALSE)

sum_squares(data, aggregate = TRUE)
}
\arguments{
\item{data}{matrix or data.frame containing numeric values.}

\item{unbiased}{flag indicating if returned empirical variance
should be unbiased (sum of squares divided by n - 1, default) or
biased (divided by n).}

\item{aggregate}{flag indicating if returned value should
be the sum of the values obtained per variable. Ignored
if \code{data} has only one column.}
}
\description{
Variance and sum of square
Tools for calculating variance and sum of squares
with a matrix or a data.frame.
}
\section{Functions}{
\itemize{
\item \code{colVars()}: Return the empirical variance for each
column of \code{data}.

\item \code{sum_squares()}: Return the sum of squares for each column
of \code{data}.

}}
\keyword{internal}
