% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3t-package.R
\docType{package}
\name{c3t-package}
\alias{c3t}
\alias{c3t-package}
\title{c3t: Clustering with Connectivity and Size Constraints}
\description{
Enable clustering with connectivity (regionalization) and/or minimum-maximum size constraints. It also provides an algorithm to attempt making a regionalization feasible (while respecting size constraints) and another to enhance a feasible solution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.lab.sspcloud.fr/ttassin/agregation-geographique-avec-continuite}
  \item Report bugs at \url{https://git.lab.sspcloud.fr/ttassin/agregation-geographique-avec-continuite/-/issues}
}

}
\author{
\strong{Maintainer}: Tangi Tassin \email{tangi.tassin@outlook.fr}

Other contributors:
\itemize{
  \item Stéphane Legleye (ensai.fr) [thesis advisor]
}

}
\keyword{internal}
