% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3t-package.R
\docType{package}
\name{c3t-package}
\alias{c3t}
\alias{c3t-package}
\title{c3t: Clustering with Connectivity and Size Constraints}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Enable clustering with connectivity (regionalization) and/or minimum-maximum size constraints. It also provides an algorithm to attempt making a regionalization feasible (while respecting size constraints) and another to enhance a feasible solution. 'c3t' stands for "Classification avec Contrainte de Connexité et de Taille", which in french means "Clustering with connecvity and or size constraints".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atero18/c3t/}
  \item \url{https://atero18.github.io/c3t/}
  \item Report bugs at \url{https://github.com/atero18/c3t/issues}
}

}
\author{
\strong{Maintainer}: Tangi Tassin \email{tangi.tassin@outlook.fr} (\href{https://orcid.org/0009-0004-5956-2910}{ORCID}) (internship)

Other contributors:
\itemize{
  \item Stéphane Legleye (\href{https://orcid.org/0000-0002-1704-2499}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
