% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat.R
\name{distMat_replace_values}
\alias{distMat_replace_values}
\alias{[<-,DistMat,numericOrMissing,numericOrMissing,numeric-method}
\alias{[<-,DistMat,matrix,missing,numeric-method}
\title{\code{DistMat} values replacement
Call the \code{AbstractSymMat} replacement method. Check that values
are all positive and if \code{NA} replace them by \code{NA_real_}}
\usage{
\S4method{[}{DistMat,numericOrMissing,numericOrMissing,numeric}(x, i, j) <- value

\S4method{[}{DistMat,matrix,missing,numeric}(x, i, j) <- value
}
\description{
\code{DistMat} values replacement
Call the \code{AbstractSymMat} replacement method. Check that values
are all positive and if \code{NA} replace them by \code{NA_real_}
}
\section{Functions}{
\itemize{
\item \code{`[`(x = DistMat, i = numericOrMissing, j = numericOrMissing) <- value}: case when \code{i} and \code{j} are vectors.
Might be missing.

\item \code{`[`(x = DistMat, i = matrix, j = missing) <- value}: case when indexes are given
by pairs in a matrix

}}
\keyword{internal}
