% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{is_connected,matrix-method}
\alias{is_connected,matrix-method}
\title{Check Connectivity}
\usage{
\S4method{is_connected}{matrix}(graph, mode = c("weak", "strong"))
}
\arguments{
\item{graph}{Contiguity matrix of the elements.
(contiguity matrix)}
}
\value{
TRUE if the set is connected, FALSE otherwise.
}
\description{
Verifies via the contiguity matrix whether a set is connected.
}
\seealso{
\code{\link[igraph:components]{igraph::is_connected()}}
}
\keyword{internal}
